fix: constant overflow when compiling for 32bit machines
on Go 1.x(x) all constants are ints by default which causing
build for 32bit machines to fail due to overflow
https://github.com/golang/go/issues/23086

+++ nodedb.go
@@ -238,7 +238,7 @@ func (ndb *nodeDB) DeleteVersionsFrom(version int64) error {
 	})
 
 	// Finally, delete the version root entries
-	ndb.traverseRange(rootKeyFormat.Key(version), rootKeyFormat.Key(math.MaxInt64), func(k, v []byte) {
+	ndb.traverseRange(rootKeyFormat.Key(version), rootKeyFormat.Key(int64(math.MaxInt64)), func(k, v []byte) {
 		if err := ndb.batch.Delete(k); err != nil {
 			panic(err)
 		}
